Move high-level functions to its own module.
This helps avoid circular imports and prevent low-level functions
accidentally use high-level functions (i.e., promoting modularization of
PyDAX).

Close #47
